Fix merge-join in query plan.
Related to #149. Fixes QueryExecutor class. It can process samples
of variable length. MergeJoinMaterializer is fixed as well. There
was a bug caused by the MergeJoinMaterializer::Range struct. This
range objects is responsible by key creation but they ignored the
template parameter. As result an output was always is the same
order.