Move roundtrip test to functests
Also, fix some stuff in the script, curl and wget don't print anything
to the stdout now and also some queries is issued to test the database
(metadata query and join query). The output of the queries is saved to
the file. It should be checked or compared with the expected output (
this file can be downloaded from S3 as well).