Add node split algorithm
Node split function added to NBTreeLeaf and NBTreeSuperblock classes.
This function splits the leaf node that contains particular timestamp
into two leafs and midifies the path from the root to the new leafs.
This functionality is supposed to be used by the test suite to check
that the nbtree-update functionality will work as intended. This
algorithm also can be modified to use in nbtree-update implementation.

The unit-tests that check the algorithm is also added.
* Tests that check two level tree split
* Tests that check three level tree split, aggregations, and backrefs