NBTreeExtentsList can split nodes randomly now
There is a small chance that the node of the tree will be split on any
append operation. The node of the extents list is chosen randomly and
the pivot point of the split is also chosen randomly (but inside the
range covered by the subtree).

The tests should go crazy on CI. The splitting functionality easy to fit
in.