Add type field into the SubtreeRef struct
The code have been using the level field to determine the type of the
node. It worked because the trees was full and all leaf nodes had
level=0 contrary to inner nodes that had level>0. This property have
been broken by the recent update that added the split method. The split
method makes the trees imbalanced and as a result not only a leaf node
can have a level field set to 0.

This commit adds new field that contains the type of the node. All the
code now checks this field to determine the type of the field instead of
looking at level field.