Fix error in NBTree
Node type of the backref has been assumed to be the same as a source
node type. When this was not the case the assert has been triggered.
The error propagation from leaf and inner nodes to extent was broken.

The check_superblock_consistency method was broken since some invariants
have changed. This commit makes the checks relevant.

Split algorithm now processes child nodes correctly and allows to make
in-place modifications to inner nodes.

The superblock iteration test was broken due to poor implementation of
the NBTreeLeaf clone c-tor.

Group-aggregate operator could return aggregate that contains data from
the outside of the query range if the query range is not the multiple of
the step.

The aggregation test has been broken because split algorithm generated
too much blocks and the test had been waiting until the critain amount
of blocks will be commited.