Improve OpenTSDB compatibility
Telnet-style OTSD protocol support improved. Version command now works.
This version had been tested with TCollector and everything seemed to
work.

The TCP server and protocol implementation has changed. Protocols
parse_next method now returns a response object. It can be empty (and
expected to be empty most of the time) but it also can contain some data
that should be sent to client without closing the connection. In older
versions the only way to send data back was through the exception. For
version command this not worked well because exception propogation
resulted in connection closing.

Version command return some arbitrary information that can be used to
determine that we dealing with Akumuli and not the OpenTSDB. It looks
like the TCollectors looks only to the first line of the response and
the second line can be anything.