Add sum and absolute functions
Old sum renamed to accumulate (that's actually a cumulative sum so
maybe it's worth to call it cusum). Absolute function just computes
an absolute value of each element of the tuple. Sum calculates the
sum of all values of the tuple and writes it back. The sample is
converted to 1-element tuple if it contained many elements.