Introduce input_log component
Component implements input log for the database. It can store
all input data in the lz4 compressed volumes (~10B per data
point) and read this data back in insert order. This component
is supposed to be used as an additional recovery mechanism.
This component uses CRoaring (as submodule) and lz4 (it's OK).
Probably, I should move away from CRoaring or package it somehow.