Fix 243
The problem only appeared when the packet size exceded 2048 bytes.
This was caused by the fact that internal receiving buffer for each
datagram was that size. This was done in assumption that the user
will want to send large amount of small packets. New version has
larger receiving buffer (64KB) that can accomodate any UDP payload.