Add CMake swith that can be used to link sqlite3 driver
The driver is the apr_dbd_sqlite3.so library. It's not installed
into the standard location so normally the linker can't find it.
Normally, it shouldn't be linked. It's loaded by the apr-dbd code
dynamically. But if we want to load libakurmuli dynamically using
the dlopen building with this option can help.