Use negative ids for event (contrary to positive ids for metrics)
Akumuli is not supposed to store different data types so sign can be
used as a flag.
Event ids are stored as negative values in the sqlite database but
in the application code they are represented using u64 (aku_ParamId)
values. Anyway, the code that have to deal with events can check if
the id corresponds to event or metric by casting to i64 and checking
the sign.